- Update Gradle and Compose versions - Create Login screen and navigation - Organize Welcome screen into its own package
