+++ app/src/main/java/ar/edu/ort/tp3_ort_2025_parcial/screen/Screens.kt
@@ -2,4 +2,5 @@ package ar.edu.ort.tp3_ort_2025_parcial.screen
 
 sealed class Screens(val screen: String) {
     data object Welcome : Screens("welcome")
+    data object Login : Screens("login")
 }
