+++ app/src/main/java/ar/edu/ort/tp3_ort_2025_parcial/screen/login/Login.kt
@@ -1,0 +1,34 @@
+package ar.edu.ort.tp3_ort_2025_parcial.screen.login
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
+
+@Composable
+fun Login(
+    navController: NavController
+) {
+    Box(
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .align(Alignment.Center),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text = "Login Screen",
+                fontSize = 30.sp,
+            )
+        }
+    }
+}
